Remove undolog.c's dependency on DSM segments.
Previously, one or more DSM segments were used to hold undo log meta
data.  This meant there was no arbitrary limit on the number of active
undo logs, but they competed with other unpredictable users of DSM
slots.  Also, it wasn't clear how to handle failure to create new DSM
segments on a hot standby.

Instead, use a fixed-sized pool of undo logs in the regular shared
memory area, with a size that is a factor of the number of max_connections.
It's possible (but hard) to run out of space on a primary, but that is
an error condition that is easier to understand and handle.

UndoRecPtrAssignRelFileNode() now expands to code that should be faster.
The previous version's call to UndoRecPtrGetTablespace() had shown up as
up to 5% of CPU time in some undo-chain-traversal-heavy workloads.

Add prevlog and status to pg_stat_undo_logs view.

Thomas Munro, RM43512