Remove xl_undolog_meta.
This was attempt #2 at dealing with undo log meta data tracking, in
crash recovery (since we need to synchronize with the most recent
checkpoint).  It involved extra parameters and retry loops all over
the place, which was not nice.  A future commit will solve the problem
with a new xlogutils.c interface to register undo logs that takes care of
everything internally.